requirements.txt
flask
paho-mqtt
psycopg2-binary
gunicorn
